backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
pymysql==1.1.0
werkzeug==3.0.4
requests==2.31.0
python-dotenv==1.0.0
dashscope==1.17.0
openai==1.12.0
numpy==1.26.4
sentence-transformers==2.3.1

